Revert "fix: fix low-severity security issues in FFI, TLV buffer, and loader"
This reverts commit 6b4547c7cd9cd5e1100ecdd481f9c925f0e90c19.

+++ framework/tlv/tlv_buffer.h
@@ -69,7 +69,7 @@ public:
         return cursor_ < total_;
     }
 
-    inline bool HasExpectBuffer(size_t expectLen) const
+    inline bool HasExpectBuffer(uint32_t expectLen) const
     {
         return total_ >= cursor_ && total_ - cursor_ >= expectLen;
     }
